:zap: :: 델리게이트 , 데이터 소스 분리
