+++ Projects/Domain/ContentDomain/Interface/Entitiy/ContentEntity.swift
@@ -9,8 +9,8 @@
 import Foundation
 
 public struct ContentEntity {
-    let id: Int
-    let title,thumbnail: String
+    public let id: Int
+    public let title,thumbnail: String
     
     public init(id: Int, title: String, thumbnail: String) {
         self.id = id
