+++ Projects/Feature/HomeFeature/Sources/ViewControllers/HomeViewController+DataSource.swift
@@ -1,0 +1,23 @@
+//
+//  HomeViewController+DataSource.swift
+//  HomeFeature
+//
+//  Created by yongbeomkwak on 11/7/23.
+//  Copyright © 2023 com. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+extension HomeViewController: UICollectionViewDataSource {
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        // 아이템 개수
+        return contentsCount
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell { // data bind
+        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: ContentsCollectionViewCell.identifier, for: indexPath)
+                as? ContentsCollectionViewCell else { return UICollectionViewCell() }
+        return cell
+    }
+}
