+++ Projects/Feature/HomeFeature/Sources/ViewControllers/HomeViewController+Output.swift
@@ -1,0 +1,26 @@
+//
+//  HomeViewController+Output.swift
+//  HomeFeature
+//
+//  Created by yongbeomkwak on 11/7/23.
+//  Copyright © 2023 com. All rights reserved.
+//
+
+import Foundation
+import RxSwift
+import UIKit
+
+extension HomeViewController {
+    func bindDataSource(output: HomeViewModel.Output) {
+        
+        Observable.zip(output.tagDataSource,output.contentDataSource)
+            .skip(1)
+            .withUnretained(self)
+            .subscribe(onNext: { (owner,dataSoruce) in
+                
+                owner.contentsCollectionView.reloadData()
+            })
+            .disposed(by: disposeBag)
+        
+    }
+}
